requirements.txt
asgiref==3.9.1
Django==5.2.4
django-admin-rangefilter==0.13.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-json-logger==4.0.0.dev0
sqlparse==0.5.3
